Turtl-core works on NW in windows =]
Finally fucking did it. Took a lot of re-wiring and careful debugging
output (which is now hardcoded into the turtl C module) but I finally
figured out what was going wrong.

when embedding ECL in node-webkit, STDOUT/STDERR seem to not just go
into a void, but to throw errors when being written to. creating dummy
vesions of *standard-output*, *error-output* fixed the issue.

now i can get down to business...